+++ ga_2_2.py
@@ -45,9 +45,9 @@ def mutacaoGene(pai, qualPai, taxa_mut, cromossomos):
       else: aux_pai.append(np.int32(1))
       arquivoTxt.write('Aconteceu mutação em pai(' + repr(qualPai) + ') : '+ repr(index+1) + '\n')
     else:
-      aux_pai.append(pai[index])
+      aux_pai.append(np.int32(pai[index]))
 
-  return aux_pai
+  return np.int32(aux_pai)
 
 # Recebe os fitness da gerações e retorna o pior e o melhor elemento
 def encontrarMelhorEPior(geracoes, melhor_filho):
